BeamOptics: PS-only or Spin+PS Push
Use the `SpinTransport` class to declare if the
element supports spin tracking. If then requested at runtime,
use the "fused" spin+PS push, otherwise use the usual PS-only
push.
