+++ src/elements/mixin/beamoptic.H
@@ -10,6 +10,8 @@
 #ifndef IMPACTX_ELEMENTS_MIXIN_BEAMOPTIC_H
 #define IMPACTX_ELEMENTS_MIXIN_BEAMOPTIC_H
 
+#include "spintransport.H"
+
 #include "particles/ImpactXParticleContainer.H"
 #include "particles/PushAll.H"
 
@@ -24,6 +26,180 @@ namespace impactx::elements::mixin
 {
 namespace detail
 {
+    /** Push a single particle through an element
+     *
+     * Note: we usually would just write a C++ lambda below in ParallelFor. But, due to restrictions
+     * in NVCC as of 11.5, we cannot write a lambda in a lambda as we also std::visit the element
+     * types of our lattice elements list.
+     *    error #3206-D: An extended __device__ lambda cannot be defined inside a generic lambda expression("operator()").
+     * Thus, we fall back to writing a C++ functor here, instead of nesting two lambdas.
+     *
+     * Nvidia bug report: 3458976
+     * Minimal demonstrator: https://cuda.godbolt.org/z/39e4q53Ye
+     *
+     * @tparam T_Element This can be a \see Drift, \see Quad, \see Sbend, etc.
+     */
+    template <typename T_Element>
+    struct PushSingleParticleSpin
+    {
+        using PType = ImpactXParticleContainer::ParticleType;
+        using ParticleTileType = amrex::ParticleTile<amrex::SoAParticle<RealSoA::nattribs, IntSoA::nattribs>,RealSoA::nattribs, IntSoA::nattribs>;
+
+        /** Constructor taking in pointers to particle data
+         *
+         * @param element the beamline element to push through
+         * @param part_x the array to the particle position (x)
+         * @param part_y the array to the particle position (y)
+         * @param part_t the array to the particle position (t)
+         * @param part_px the array to the particle momentum (x)
+         * @param part_py the array to the particle momentum (y)
+         * @param part_pt the array to the particle momentum (t)
+         * @param part_sx the array to the particle spin (x)
+         * @param part_sy the array to the particle spin (y)
+         * @param part_sz the array to the particle spin (z)
+         * @param part_idcpu the array to the particle global index
+         * @param ref_part the struct containing the reference particle
+         */
+        PushSingleParticleSpin (T_Element element,
+                            amrex::ParticleReal* AMREX_RESTRICT part_x,
+                            amrex::ParticleReal* AMREX_RESTRICT part_y,
+                            amrex::ParticleReal* AMREX_RESTRICT part_t,
+                            amrex::ParticleReal* AMREX_RESTRICT part_px,
+                            amrex::ParticleReal* AMREX_RESTRICT part_py,
+                            amrex::ParticleReal* AMREX_RESTRICT part_pt,
+                            amrex::ParticleReal* AMREX_RESTRICT part_sx,
+                            amrex::ParticleReal* AMREX_RESTRICT part_sy,
+                            amrex::ParticleReal* AMREX_RESTRICT part_sz,
+                            uint64_t* AMREX_RESTRICT part_idcpu,
+                            RefPart ref_part)
+            : m_element(std::move(element)),
+              m_part_x(part_x), m_part_y(part_y), m_part_t(part_t),
+              m_part_px(part_px), m_part_py(part_py), m_part_pt(part_pt),
+              m_part_sx(part_sx), m_part_sy(part_sy), m_part_sz(part_sz),
+              m_part_idcpu(part_idcpu),
+              m_ref_part(std::move(ref_part))
+        {
+            // pre-compute and cache constants that are independent of the
+            // individually tracked particle
+            m_element.compute_constants(m_ref_part);
+        }
+
+        PushSingleParticleSpin () = delete;
+        PushSingleParticleSpin (PushSingleParticleSpin const &) = default;
+        PushSingleParticleSpin (PushSingleParticleSpin &&) = default;
+        ~PushSingleParticleSpin () = default;
+
+        /** Push a single particle through an element
+         *
+         * @param i particle index in the current box
+         */
+        AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+        void
+        operator() (int i) const
+        {
+            // access SoA Real data
+            // note: an optimizing compiler will eliminate loads of unused parameters
+            amrex::ParticleReal & AMREX_RESTRICT x = m_part_x[i];
+            amrex::ParticleReal & AMREX_RESTRICT y = m_part_y[i];
+            amrex::ParticleReal & AMREX_RESTRICT t = m_part_t[i];
+            amrex::ParticleReal & AMREX_RESTRICT px = m_part_px[i];
+            amrex::ParticleReal & AMREX_RESTRICT py = m_part_py[i];
+            amrex::ParticleReal & AMREX_RESTRICT pt = m_part_pt[i];
+            amrex::ParticleReal & AMREX_RESTRICT sx = m_part_sx[i];
+            amrex::ParticleReal & AMREX_RESTRICT sy = m_part_sy[i];
+            amrex::ParticleReal & AMREX_RESTRICT sz = m_part_sz[i];
+            uint64_t & AMREX_RESTRICT idcpu = m_part_idcpu[i];
+
+            // push spin & phase space through element
+            m_element.spin_and_phasespace_push(x, y, t, px, py, pt, idcpu, m_ref_part, sx, sy, sz);
+        }
+
+#ifdef AMREX_USE_SIMD
+        /** Push N SIMD particles through an element
+         *
+         * @param i particle index in the current box
+         */
+        template<int SIMD_WIDTH>
+        AMREX_GPU_DEVICE AMREX_FORCE_INLINE
+        void
+        operator() (amrex::SIMDindex<SIMD_WIDTH, int> idx) const
+        {
+            using namespace amrex::simd;
+
+            int const i = idx.index;
+
+            // uninitialized memory
+            SIMDParticleReal<SIMD_WIDTH> part_x;
+            SIMDParticleReal<SIMD_WIDTH> part_y;
+            SIMDParticleReal<SIMD_WIDTH> part_t;
+            SIMDParticleReal<SIMD_WIDTH> part_px;
+            SIMDParticleReal<SIMD_WIDTH> part_py;
+            SIMDParticleReal<SIMD_WIDTH> part_pt;
+            SIMDParticleReal<SIMD_WIDTH> part_sx;
+            SIMDParticleReal<SIMD_WIDTH> part_sy;
+            SIMDParticleReal<SIMD_WIDTH> part_sz;
+            SIMDIdCpu<decltype(part_x)> part_idcpu;
+
+            // initialize vector registers
+            // TODO stdx::vector_aligned needs alignment guarantees
+            //      https://github.com/AMReX-Codes/amrex/issues/4592
+            //      https://en.cppreference.com/w/cpp/experimental/simd/simd/copy_from
+            part_x.copy_from(&m_part_x[i], stdx::element_aligned);
+            part_y.copy_from(&m_part_y[i], stdx::element_aligned);
+            part_t.copy_from(&m_part_t[i], stdx::element_aligned);
+            part_px.copy_from(&m_part_px[i], stdx::element_aligned);
+            part_py.copy_from(&m_part_py[i], stdx::element_aligned);
+            part_pt.copy_from(&m_part_pt[i], stdx::element_aligned);
+            part_sx.copy_from(&m_part_sx[i], stdx::element_aligned);
+            part_sy.copy_from(&m_part_sy[i], stdx::element_aligned);
+            part_sz.copy_from(&m_part_sz[i], stdx::element_aligned);
+            part_idcpu.copy_from(&m_part_idcpu[i], stdx::element_aligned);
+
+            // push spin & phase space through element
+            m_element.spin_and_phasespace_push(part_x, part_y, part_t, part_px, part_py, part_pt, part_sx, part_sy, part_sz, part_idcpu, m_ref_part);
+
+            // write back to memory
+            // TODO stdx::vector_aligned needs alignment guarantees
+            //      https://github.com/AMReX-Codes/amrex/issues/4592
+            //      https://en.cppreference.com/w/cpp/experimental/simd/simd/copy_from
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 0))
+                part_x.copy_to(&m_part_x[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 1))
+                part_y.copy_to(&m_part_y[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 2))
+                part_t.copy_to(&m_part_t[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 3))
+                part_px.copy_to(&m_part_px[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 4))
+                part_py.copy_to(&m_part_py[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 5))
+                part_pt.copy_to(&m_part_pt[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 6))
+                part_sx.copy_to(&m_part_sx[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 7))
+                part_sy.copy_to(&m_part_sy[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 8))
+                part_sz.copy_to(&m_part_sz[i], stdx::element_aligned);
+            if constexpr (is_nth_arg_non_const(&T_Element::template operator()<decltype(part_x), decltype(part_idcpu)>, 9))
+                part_idcpu.copy_to(&m_part_idcpu[i], stdx::element_aligned);
+        }
+#endif
+
+    private:
+        T_Element m_element;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_x;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_y;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_t;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_px;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_py;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_pt;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_sx;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_sy;
+        amrex::ParticleReal* const AMREX_RESTRICT m_part_sz;
+        uint64_t* const AMREX_RESTRICT m_part_idcpu;
+        RefPart m_ref_part;
+    };
+
     /** Push a single particle through an element
      *
      * Note: we usually would just write a C++ lambda below in ParallelFor. But, due to restrictions
@@ -179,7 +355,8 @@ namespace detail
     void push_all_particles (
         ImpactXParticleContainer::iterator & pti,
         RefPart & AMREX_RESTRICT ref_part,
-        T_Element & element
+        T_Element & element,
+        bool spin /* = false */
     ) {
         const int np = pti.numParticles();
 
@@ -191,20 +368,43 @@ namespace detail
         amrex::ParticleReal* const AMREX_RESTRICT part_px = soa_real[RealSoA::px].dataPtr();
         amrex::ParticleReal* const AMREX_RESTRICT part_py = soa_real[RealSoA::py].dataPtr();
         amrex::ParticleReal* const AMREX_RESTRICT part_pt = soa_real[RealSoA::pt].dataPtr();
+        amrex::ParticleReal* const AMREX_RESTRICT part_sx = soa_real[RealSoA::sx].dataPtr();
+        amrex::ParticleReal* const AMREX_RESTRICT part_sy = soa_real[RealSoA::sy].dataPtr();
+        amrex::ParticleReal* const AMREX_RESTRICT part_sz = soa_real[RealSoA::sz].dataPtr();
 
         uint64_t* const AMREX_RESTRICT part_idcpu = pti.GetStructOfArrays().GetIdCPUData().dataPtr();
 
-        detail::PushSingleParticle<T_Element> const pushSingleParticle(
-            element, part_x, part_y, part_t, part_px, part_py, part_pt, part_idcpu, ref_part);
+        if (spin) {
+            if constexpr (std::is_base_of_v<mixin::SpinTransport, std::decay_t<T_Element>>) {
+                detail::PushSingleParticleSpin<T_Element> const pushSingleParticle(
+                    element, part_x, part_y, part_t, part_px, part_py, part_pt, part_sx, part_sy, part_sz, part_idcpu, ref_part);
 
-        // loop over beam particles in the box
+                // loop over beam particles in the box
 #ifdef AMREX_USE_SIMD
-        if constexpr (amrex::simd::is_vectorized<T_Element>) {
-            amrex::ParallelForSIMD<T_Element::simd_width>(np, pushSingleParticle);  // TODO: test 2,4,8, ... and handle remainder
-        } else
+                if constexpr (amrex::simd::is_vectorized<T_Element>) {
+                    amrex::ParallelForSIMD<T_Element::simd_width>(np, pushSingleParticle);  // TODO: test 2,4,8, ... and handle remainder
+                } else
 #endif
-        {
-            amrex::ParallelFor(np, pushSingleParticle);
+                {
+                    amrex::ParallelFor(np, pushSingleParticle);
+                }
+            } else {
+                throw std::runtime_error("Spin transport requested but element ... does not implement the `SpinTransport` interface class!");
+            }
+        }
+        else {
+            detail::PushSingleParticle<T_Element> const pushSingleParticle(
+                element, part_x, part_y, part_t, part_px, part_py, part_pt, part_idcpu, ref_part);
+
+            // loop over beam particles in the box
+#ifdef AMREX_USE_SIMD
+            if constexpr (amrex::simd::is_vectorized<T_Element>) {
+                amrex::ParallelForSIMD<T_Element::simd_width>(np, pushSingleParticle);  // TODO: test 2,4,8, ... and handle remainder
+            } else
+#endif
+            {
+                amrex::ParallelFor(np, pushSingleParticle);
+            }
         }
     }
 } // namespace detail
@@ -244,10 +444,12 @@ namespace detail
          *
          * @param[in] pti particle iterator for a current tile or box.
          * @param[in] ref_part reference particle
+         * @param[in] spin whether to perform spin tracking
          */
          void operator() (
             ImpactXParticleContainer::iterator & pti,
-            RefPart & AMREX_RESTRICT ref_part
+            RefPart & AMREX_RESTRICT ref_part,
+            bool spin /* = false */
          )
          {
             static_assert(
@@ -256,7 +458,7 @@ namespace detail
             );
 
             T_Element& element = *static_cast<T_Element*>(this);
-            detail::push_all_particles<T_Element>(pti, ref_part, element);
+            detail::push_all_particles<T_Element>(pti, ref_part, element, spin);
          }
     };
