+++ src/elements/mixin/spintransport.H
@@ -23,9 +23,13 @@
 
 namespace impactx::elements::mixin
 {
-    /** This is a helper class for applying a spin map generated by a three-component vector
-     *  lambda, whose magnitude defines the angle of rotation and whose direction defines
-     *  the axis of rotation.
+    /** This is a helper class for Spin Transport
+     *
+     * If an element derives from this class, then it must implement the `spin_and_phasespace_push` method.
+     *
+     * It also provides a helper function `rotate_spin` for applying a spin map generated
+     * by a three-component vector lambda, whose magnitude defines the angle of rotation
+     * and whose direction defines the axis of rotation.
      */
     struct SpinTransport
     {
